requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
motor==3.3.1
pydantic==2.5.2
jinja2==3.1.2
python-multipart==0.0.6
SpeechRecognition==3.14.1
pyttsx3==2.90
openai-whisper==20240930
soundfile==0.12.1
numpy==1.26.2
torch==2.1.2
numba==0.58.1
tqdm==4.66.1
more-itertools==10.1.0
tiktoken==0.5.2